Controls Tweaked
I created a new script which uses Jet Movement instead of Free Movement.
Jet Movement means that the ship is always moving forward, and only
thrusters can change your speed. Looking around with the mouse works
like usual. I did notice that if the ship collides with a planet, the
movement of the ship is permanently messed up for that game. I will look
into that and fix it in a future update.