Fixed ship rotation and collision
The ship no longer bounces of planets. We can handle planet collision
with scripts instead. I also fixed rotation of the ship to be based on
its current axis, not the world axis. This leads to much better
controlling of the ship when making wild space turns.

As a side note, I also made the bullet projectile move faster like an
actual bullet or laser.