Standalone Camera
The camera is no longer a child of the player. This will allow us to
modify the camera to behave properly for aerial maneuvers later. The new
camera script is written in JavaScript. (If you need to modify the
camera, I hope you know JavaScript. Otherwise, you can talk to me, and I
will happily do it for you!)