Created energy system
The ship now has a finite amount of energy. This energy is spent
boosting, braking, and performing maneuvers. After a few seconds of not
using energy, it will begin to rise. I am not sure how to change the
font for my GUI, so it is simply black text. I'll ask someone how to fix
that later.