Powerup added
Right now, I have only created a shield powerup. And at the moment,
there is no graphical change for the shield. I attempted to get the
purple shield model working, but it never showed up. In any case, I am
heading to bed. I will make other powerups tomorrow. Also, if we decide
powerups are not going to work in the long run, we can scrap this branch
or save it for later.