Allowed for First-Person Camera and for Alternate Control Schemes
Pressing "P" will allow the player to switch between third-person and
first-person view.

Pressing "C" will change the controls. The horizontal mouse axis
switches with A and D. This means that the horizontal mouse will rotate
around the up axis instead of the forward axis, while A and D will
rotate around the forward axis instead of the up axis.