Finalized Powerups system
The powerups can now be set to have a certain number within the arena at
one time. I have set the default to 5 of each, since the arena is so
huge. The powerups also glow their color, and will teleport to another
area of the arena after 30 seconds. They also spawn randomly instead of
right at the origin.